+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # 1.8
 
+ - Add WildStacker integration
+
 # 1.7
 
  - Citizens NPCs are ignored by spawn rules (by default, use target_npc: true to change)